Added a version of pureMixture with function speciesData
This enabled the basic library to be built since previously the function name we needed was not available in the mixture classes in basic but in reactionThermo only. This is probably not necessary but I wanted to keep the libs seperate and analogous files in their proper lib folders. Im not sure how to include the .C file from basic in the reactionThermo lib more intelegently than just compiling the basic lib and including the lnInclude folder.