Reduce calls to thermo.rho() in CellState
Previously we would call thermo.rho() to update the density in each cell.
I now realize that is not a stored field but is instead calculated from
(psi * P). To avoid doing that I will just store the entire field once at
the beginning of each soot updateSources() call. Since we treat the thermo
fields as frozen anyway it makes no difference, other than to increase speed.

It might be a good idea in the future to do a similar thing with pressure
and temperature, even though they are stored within the thermo class
(i.e. not  calculated at each access call as density is).