Set the sootVolume_ boundary field to zero
One of the bugs was that the boundary field for sootVolume_ was never
set. This led to pretty crazy density calculations from garbage
soot volume numbers. For now I am forcing the boundary fields to be zero
and we therefore fall back on rho = P_ * psi_ instead of the full
multiphase treatment for the boundaries.

I think the only reasonable alternative would be to somehow interpolate all
of the volume fraction calculations to the boundaries.