Clean file structure for multiphase density fix
I originally wanted to maintain a seperate sootBasic and sootReactionThermo
in analogy to the OF directory structure. It is a pain because all we
really need from basic is sootHePsiThermo.C/.H. I just moved them into the
sootReactionThermo directory and removed the Make links to the old
sootBasic directory.