Renames EulerImplicitSystem to TwoEquationSoot
When first developed this library was going to handle the integration
of the soot variables in an implicit manner and I thought that
the system could be reused. The library instead involved into a soot
specific model. This name is much more diescriptive of the model and
a 2 equation soot model is really all that it adds.