Places physical limit on tar breakdown amount in PCRM
Previously the amount of tar breakdown products being released was not
limited by the actual amount of primary tar. I.e. we would, potentially,
release tar breakdown products when there hadn't been the equivalent mass
of tar produced yet. Now a check is performed and the mass of products
released is limited to the amount of tar available for breakdown.