Finished change to hashtable for local specie tracking in SootTarModel
I am copying TwoEquationSoot to make SootTarModel and trying to correct
things I wish I had done differently along the way (in the future I would
like to generalize this a make a base class that they share since they are
so similar but we need to just run now). One of the biggest changes is using
a hashtable of a list, one entry for each specie, as a member to the main class.
This then is reset for each cell.
Previously (i.e. in TwoES) we used a list that was recreated in each cell.
That meant that we needed to have a set number of species (this was hardcoded)
and just remember which names corresponded to which entry. Now we can avoid
using a list and memorizing names and just use this class member hash table
speciesY_. Then just reset it from CellState for each cell as we loop through.

This importantly allows some variation in the tar breakdown product species too
those are now read at runtime from the contstant/tarBreakdown dict file.