Completes the SootTarModel class
Finished the conversion from TwoEquationSoot. I made some improvements that
I wish I had though of the last time but there is still a lot to be done. I
really think in the future a shared baseclass between the two is the best
option.

It has not been tested at all at this point. I will start that this afternoon.