Creates new CellState constructor to work with SootTarModel
The original CellState constructor had the TwoEquationSoot model species
hardcoded into it's tables. In other words it tracked the mass fractions
and partial pressrues of the species that are needed in TES. The species
needed in STM are different from those of TES.

To fix this I added a new constructor that is more general. It takes two
wordLists in addition to the parameters already used. One of the lists
contains the species names that CellState should track the mass fractions
for and the other contains names for partial pressures to be tracked.

It is just a better and more general design all around and I would eventually
like to migrate TES to use it too. I haven't done that yet though because
at this point we just want to get running.