Prevents sourceY and sourceN functions in TwoEquationSoot from resetting the source fields after recording
Previously when the source functions set a returned object equal to the local source
field that local source field was zeroed afterwards. I'm not really sure why I did that
in the first place and it has now caused two problems when we call the source
functions repeatedly in a solver the second call returns zero. So in this commit I
removed the zeroing out.

zeroing out of that source field