Prevents zeroing of source field upon call to sourceY in SootTarModel
Just as done in the previous commit, and for the same reasons, this will
remove the zeroing out of the species source fields following their
return to a solver in sourceY.

Also I removed the sourceNs function from SootTarModel. It was an artifact from
TwoEquationSoot and is not relevant for this model.