Fixes an issue with the debugging rate write-out feature
We have been creating rate volScalarFields in the solvers and passing them
to the soot model to get an idea of which of the five reactions considered
there are having the largest effect. This is useful for debugging, if we have
a sudden temperature/species increase we can look at the rates to diagnose the issue.

This has been a feature all along and previously they are updated every time the
reaction rates are calculated. This is a problem when multiple sub-dt time steps are
used because e.g. O2 can be entirely consumed in the first time step (with a high
soot oxidation rate) but once the O2 supply in a cell is exhausted the soot
oxidation rate will become zero. Then in future sub-dt time steps we record a
zero oxidation rate when in fact it was initially quite large.

With this commit we still record the reaction rates at every time step but now
we have a second record. That second record is set equal to the sum of the
sub-dt reaction rate multiplied with the sub-dt itself. At the end of the total dt
we divide that sum for each rate by the overall dt. This effecticely averages the
reaction rates over the total time step dt and provides a much more useful quantity
for debugging purposes.