Fixes serious bug in TwoEquationSoot model explicit step function
I had a typo in the explcit step for species CO. Look at the diff and its clear
what I mistyped. The effect, in an energetic sense, is that the H2O reaction became
exothermic, and highly so. Every reaction produced 3 moles of CO when only one
should have been produced. Since CO has a negative enthalpy of formation it made
the reaction exothermic since all we do for the enthalpy source is sum the
product of species sources with their h_f.

I believe that this may fully explain the troubles we have had recently with excessive
temperatures where soot chemistry is taking place.