Adds new devolatilization model to be used with SootTarFoam solver
Previsouly I had been running the STF solver with the built-in singleKineticRate
devolatilization model from OF. I realized that there was no reason that the PCCL
rate constants (A and E) should work within that framework today. I since have
determined that they are not in fact directly applicable and we need to have a
custom devolatilization model for their use.

The problem is not that PCCL describes daf weight loss, those effects cancel. The
issue is that the rate in PCCL is determined with the fraction yielded from the initial
mass while in OF the rate is determined with the current mass of the volatile remaining
which is a function of the current particle mass and not the original particle mass.

Using the built-in model with STF results in very reasonable rates but they are still
not quite correct I now realize.

** Currently the new model 'PcSforDevolatilization' is functionaly identical to
   the built-in one that it's copied from. I just created the library and compiled
   it to get things going tonight. Tomorrow I will be back to actually implement
   the changes.