Adds new devolatilization model for SootTarFoam
I discovered while implementing the PCCL version of this model that the built-in
OF model does the same thing and we can take the rate constants directly from PCCL
and plug them into OF. I.e. my comments on the last commit were incorrect. So I have
not changed anything on that front.

One caveat is that we may still need to convert the units of the activation energy if
we wish to utilize the rates generated by PCCL. I made that simpler and less confusing
in this model. Now the user is forced to select
a unit for the AE, either PCCL -> kcal/(mol K) or  OF -> J/(kmol K). This way the user
is forced to at least consider the issue.

I also implemented the analytical integration here rather than the explicit method
used in the built-in model. It doesn't do too much in my experience but it doesn't
cost any more to run (its the same number of operations actually) and in cases of
high rates it may help stability somewhat.