Adds rate fields for the SootTarModel
Just as we have for the soot model. We track oxidation, cracking and soot formation
rates in each cell and store it in these new fields to see in post-processing.

I also added a field to track the number of explicit steps taken. This should be equal
to the number of substeps specified by the caller, the default is three. If its more
then it's not necessarily a problem it just means that there has been some over
consumption.