Forces user to use Pyrene as tar surrogate in TT model
Previously the user could use any tar surrogate they wished in their
combustion model. This almost works but the thermo database is used to fine
the MW in the TT model tar breakdown functions, which are also hard-coded
to have Pyrene as tar. If the napthalene is used for example (it's smaller
than Pyrene) we will have an underconsumption of tar and an over production
of tar breakdown products, most notably soot.

I also made a small change in a comment in the PC model, I added a TODO to
a potential bug there.