Fixes serious erroneous soot formation coeff in STM
I made a mistake in my calculation of the stoichiometric coeffcients for the
soot formation reaction in the SootTarModel. The end result was that soot was
being over produced. The mass was not even balanced so we lost 1 [g] of tar and
gained something like 2 [g] of soot. This explains the issues we have been
seeing I hope. At least it does in combination with the other problems that
previous commits have addressed.