Fixes error in universal gas constant calculation in ASFORRate devol
I forgot to add the activation energy units variable to the
copy constructor. That constructor isnt used in GTS runs but
it is used when you run in LTS. This led to the local R being set
to zero and throwing a sigFPE.

It was a stupid way to handle it anyway. I had been setting the R
everytime the calculate function was called. Now we just set it
once in the constructor (the normal constructor). And of course I
successfully added it to the copy constructor this time.