Fixes rate reporting bug in SootTarFoam and derivatives
In the sootTarFoam model there are scalarFields that are references to the
corresponding fields in the solvers that are passed during construction of
the main object.
Those fields were not being zeroed out every time step and were only added
to. So it eventually becomes an averaged quantity when what we really want is
an instantaneous snapshot. So I added the zeroRateRefs() function that is
called at the beginning of each gas phase time step to reset those rates.

I also applied this fix to the PCFSootTarModel library.