Fixes the multiphase density calculation issue
We have already implemented the correction that includes the soot
density contribution in a multiphase fashion. The implementation that
I had before averaged the compressibilities of all gas phase species,
excluding soot, to subsequently use the IGL in calculating the density.
That gas density is then used to calculate the true multiphase density
(see previous commits).

In that original implementation I forgot to renomalize the remaining gas
phase species mass fractions once soot is excluded. Since the soot
mass fraction can be quite high on occasions this error can have a
real effect. I have fixed that error now and the gas phase density is calcuated
with mass fractions that sum to one. It is like the soot is not there at all,
which is just what we want.